refactor(sell): refactor sell form data into mapper
